backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.1.0
django-filter==23.1
webcolors==1.12
psycopg2-binary==2.9.5
Pillow==9.3.0
pytest==7.1.3
pytest-django==4.5.2
PyYAML==6.0
gunicorn==20.1.0
python-dotenv==1.1.0